+++ src/sys/hw/_.rs
@@ -1,7 +1,7 @@
 // devela/src/sys/hw/_.rs
 //
 #![doc = crate::_DOC_SYS_HW!()] // public
-#![doc = crate::_doc!(modules: crate::sys; hw)] //
+#![doc = crate::_doc!(modules: crate::sys; hw: mcu)] //
 #![doc = crate::_doc!(flat:"sys")]
 #![doc = crate::_doc!(hr)]
 //!