+++ src/sys/hw/mcu/_.rs
@@ -1,6 +1,9 @@
 // devela/sys/hw/mcu/_.rs
 //
-//! Microcontroller units and bare-metal execution substrates.
+#![doc = crate::_DOC_SYS_HW_MCU!()] // public
+#![doc = crate::_doc!(modules: crate::sys::hw; mcu: avr, board)] // esp32
+#![doc = crate::_doc!(flat:"sys")]
+#![doc = crate::_doc!(hr)]
 //!
 //! This module contains chip-specific foundations
 //! for programs that execute directly on microcontrollers.
@@ -14,20 +17,23 @@
 
 crate::mods_in! {
     pub mod_ avr;
+    pub mod_ board;
     // pub mod_ esp32;
 }
 crate::mods_out! { // _pub_mods, _reexports
     _pub_mods {
         pub use super::{
             avr::_all::*,
+            board::_all::*,
             // esp32::_all::*,
         };
     }
     _reexports {
         #[doc(inline)]
         pub use super::{
-            avr::{AvrPort, AvrReg8}, // Atmega328p
-            // esp32::{},
+            avr::{Atmega328p, AvrPort, AvrReg8, AvrUsart},
+            // board::{},
+            // esp32::{Esp32C3, EspReg32},
         };
     }
 }